requirements.txt
coverage==6.4.1
Django==4.0.5
djangorestframework==3.13.1
djangorestframework_simplejwt==5.2.0
django-nose==1.4.7
psycopg2-binary==2.9.3
python-dotenv==0.20.0

